Update page and locator data, splittable between headpin and katello
Can now specify that pages or locators are for headpin, katello, or either.  You can switch servers at the repl "live" and it should 'just work' because it caches per-url whether it is a katello or headpin server (asking via api if it hasn't asked that server before).  The only way this will fail is if a hostname is reinstalled and changed to a different deployment while the repl is still up.  If that happens you can just do (require 'katello.rest :reload) to reset the cache.